requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
dnspython==2.8.0
drf-yasg==1.21.11
email-validator==2.3.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jwt==1.4.0
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.23
pytz==2025.2
pyyaml==6.0.3
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
